Move siren schema tests to siren_schema.ts
