+++ api_tests/tests/rfc003_schema.ts
@@ -3,12 +3,7 @@ import * as sirenJsonSchema from "../siren.schema.json";
 import * as swapPropertiesJsonSchema from "../swap.schema.json";
 import { twoActorTest } from "../src/actor_test";
 import { createDefaultSwapRequest, DEFAULT_ALPHA } from "../src/utils";
-import {
-    Action,
-    EmbeddedRepresentationSubEntity,
-    Entity,
-    Link,
-} from "comit-sdk";
+import { Action, EmbeddedRepresentationSubEntity, Link } from "comit-sdk";
 import { extendSchemaMatcher } from "../src/schema_matcher";
 
 extendSchemaMatcher();
@@ -17,18 +12,6 @@ extendSchemaMatcher();
 // RFC003 schema tests                          //
 // ******************************************** //
 
-async function assertValidSirenDocument(swapsEntity: Entity, alice: Actor) {
-    const selfLink = swapsEntity.links.find((link: Link) =>
-        link.rel.includes("self")
-    ).href;
-
-    const swapResponse = await alice.cnd.fetch(selfLink);
-    const swapEntity = swapResponse.data as Entity;
-
-    expect(swapEntity).toMatchSchema(sirenJsonSchema);
-    expect(swapEntity.properties).toMatchSchema(swapPropertiesJsonSchema);
-}
-
 describe("Rfc003 schema tests", () => {
     it(
         "get-all-swaps-is-valid-siren",
@@ -36,31 +19,6 @@ describe("Rfc003 schema tests", () => {
             const res = await alice.cnd.fetch("/swaps");
 
             expect(res.data).toMatchSchema(sirenJsonSchema);
-        })
-    );
-
-    it(
-        "get-single-swap-is-valid-siren",
-        twoActorTest(async ({ alice, bob }) => {
-            // Alice send swap request to Bob
-            await alice.cnd.postSwap(await createDefaultSwapRequest(bob));
-
-            const aliceSwapEntity = await alice
-                .pollCndUntil("/swaps", (body) => body.entities.length > 0)
-                .then(
-                    (body) =>
-                        body.entities[0] as EmbeddedRepresentationSubEntity
-                );
-
-            await assertValidSirenDocument(aliceSwapEntity, alice);
-
-            const bobsSwapEntity = await bob
-                .pollCndUntil("/swaps", (body) => body.entities.length > 0)
-                .then(
-                    (body) =>
-                        body.entities[0] as EmbeddedRepresentationSubEntity
-                );
-            await assertValidSirenDocument(bobsSwapEntity, bob);
         })
     );
